requirements.txt
altair==5.5.0
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
JustWatch==0.5.1
kafka-python==2.2.15
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.46.0
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrends==4.9.2
pytz==2025.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
